requirements.txt
django==1.8.5
MySQL-python==1.2.5
Pillow==3.2.0( or PIL)